+++ src/main/java/com/uwetrottmann/tmdb2/entities/WatchProvidersResults.java
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2025 Uwe Trottmann
+
+package com.uwetrottmann.tmdb2.entities;
+
+
+import java.util.List;
+
+public class WatchProvidersResults {
+
+    public List<WatchProviders.WatchProvider> results;
+
+}